pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "agentbahn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=47.0.0",
    "django>=6.0.4",
    "django-click>=2.5.0",
    "django-ninja",
    "dspy>=3.1.3",
    "dspy-codex",
    "httpx",
    "loguru>=0.7.3",
    "mlflow-skinny>=3.11.1",
    "pandas>=3.0.2",
    "pyright>=1.1.408",
    "python-dotenv>=1.0.1",
    "textual>=8.2.4",
]

[project.scripts]
agentbahn-tui = "agentbahn_tui.cli:main"

[dependency-groups]
dev = [
    "django-extensions>=4.1",
    "django-stubs>=6.0.3",
    "model-bakery>=1.23.4",
    "ptpython>=3.0.32",
    "pyright>=1.1.408",
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
    "ruff>=0.15.11",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "agentbahn.settings"

[tool.ruff]
target-version = "py314"

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
include = ["agentbahn", "agentbahn_tui"]
pythonVersion = "3.14"
typeCheckingMode = "standard"
reportMissingImports = "warning"
venvPath = "."
venv = ".venv"

[tool.django-stubs]
django_settings_module = "agentbahn.settings"

[tool.setuptools.packages.find]
include = ["agentbahn", "agentbahn.*", "agentbahn_tui", "agentbahn_tui.*"]

[tool.uv.sources]
dspy-codex = { path = "../dspy_codex", editable = true }
